display only time picker while editing recurrence rules
