+++ js/app/i18n/locales/en.json
@@ -700,8 +700,8 @@
         "DELIVERY_FORM_SHOW_LESS" : "Show less",
         "DELIVERY_PICKUP": "Pickup",
         "DELIVERY_DROPOFF": "Dropoff",
-        "DELIVERY_FORM_PICKUP_HOUR": "Pickup hour",
-        "DELIVERY_FORM_DROPOFF_HOUR": "Dropoff hour",
+        "DELIVERY_FORM_PICKUP_HOUR": "Pickup window",
+        "DELIVERY_FORM_DROPOFF_HOUR": "Dropoff window",
         "DELIVERY_FORM_NEW_PRICE": "New price",
         "DELIVERY_FORM_OLD_PRICE": "Old price",
         "DELIVERY_FORM_TRACKING_LINK": "Tracking link",
